requirements.txt
# Framework FastAPI
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Base de données
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Authentification et sécurité
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# Client HTTP
httpx==0.26.0

# Traitement de texte et markdown
markdown==3.5.2
pyyaml==6.0.1

# Utilitaires
python-dotenv==1.0.1
tenacity==8.2.3

# Tests
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
aiosqlite==0.19.0